+++ src/components/CommentList.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
 import { ref } from 'vue'
-import { formatDate } from '../utils/formatDate'
-import { useMemiBoardComments } from '../composables/useMemiBoardComments'
-import { useMemiBoardAuth } from '../composables/useMemiBoardAuth'
+import { formatDate } from 'memi-board'
+import { useMemiBoardComments } from 'memi-board'
+import { useMemiBoardAuth } from 'memi-board'
 
 const props = defineProps<{ postId: string }>()